backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
alembic==1.13.1
aiomysql==0.2.0
PyMySQL==1.1.1
greenlet>=3.0.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.9
websockets==12.0
httpx==0.27.0
python-dotenv==1.0.1
